requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
Pillow==8.1.0
psycopg2==2.8.6
PyJWT==2.0.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3
